Bugfix/fix scoring deploy issues (#262)
* fixed migration error if column exists

* acounted for missing last_status_changed for lead

* fix: include missed class import in appointmentcontroller

* fix: fix typo on leadController, add styles for datepicker in console

+++ database/migrations/2019_12_11_002822_add_outcome_column_to_recipients_table.php
@@ -14,7 +14,9 @@ class AddOutcomeColumnToRecipientsTable extends Migration
     public function up()
     {
         Schema::table('recipients', function (Blueprint $table) {
-            $table->string("outcome", 50)->nullable()->after("last_status_changed_at");
+            if (! Schema::hasColumn('recipients', 'outcome')) {
+                $table->string("outcome", 50)->nullable()->after("last_status_changed_at");
+            }
         });
     }
 
@@ -26,7 +28,9 @@ class AddOutcomeColumnToRecipientsTable extends Migration
     public function down()
     {
         Schema::table('recipients', function (Blueprint $table) {
-            $table->dropColumn("outcome");
+            if (Schema::hasColumn('recipients', 'outcome')) {
+                $table->dropColumn("outcome");
+            }
         });
     }
 }
